+++ src/modules/create-orders/create-orders.controller.ts
@@ -1,18 +1,21 @@
-import { Controller, Get, Query } from '@nestjs/common';
+import { Body, Controller, Get, Post, UseGuards } from '@nestjs/common';
+import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
+import { Role } from 'src/auth/decorators/roles.decorator';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
+import { RolesGuard } from 'src/auth/guards/roles.guard';
 import { CreateOrdersService } from './create-orders.service';
+import { PedidoDTO } from './dtos/pedido.dto';
 
 @Controller('create-orders')
-//@UseGuards(JwtAuthGuard, RolesGuard)
+@UseGuards(JwtAuthGuard, RolesGuard)
 export class CreateOrdersController {
     constructor(private readonly CreateOrderService: CreateOrdersService) { }
 
     @Get()
     async GetProductsOrders(
-        @Query('codven') codven?: string
-        // @CurrentUser('role') role: string,
-        // @CurrentUser('codven') codven: string,
+        @CurrentUser('role') role: string,
+        @CurrentUser('codven') codven: string,
     ) {
-
         return await this.CreateOrderService.GetProductsOrders(codven);
     }
 
@@ -24,8 +27,8 @@ export class CreateOrdersController {
 
     @Get("iva")
     async GetIVA() {
-   
-      return await this.CreateOrderService.GetIVA(); 
+
+        return await this.CreateOrderService.GetIVA();
 
     }
 
@@ -33,5 +36,16 @@ export class CreateOrdersController {
     async GetConditions() {
 
         return await this.CreateOrderService.GetConditions();
+    }
+
+    @Role('2', '3', '5')
+    @Post()
+    async CreateOrder(
+        @Body() createOrder: PedidoDTO,
+        @CurrentUser('codven') codven: string,) {
+
+        const factNumber = await this.CreateOrderService.InsertOrder(createOrder, codven);
+        return { factNumber };
+
     }
 }