+++ src/modules/create-orders/create-orders.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from 'src/database/database.module';
+import GoalsService from '../goals/goals.service';
 import { CreateOrdersController } from './create-orders.controller';
 import { CreateOrdersService } from './create-orders.service';
 
 @Module({
 
     imports:[DatabaseModule],
-    providers:[CreateOrdersService],
+    providers:[CreateOrdersService, GoalsService],
     controllers:[CreateOrdersController]
 })
 export class CreateOrdersModule {}